MoodyBeatsRecommenderAPI/requirements.txt
boto==2.49.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.20
django-cors-headers==2.0.0
django-storages==1.6.5
djangorestframework==3.8.2
hellowebapp-deploy==1.0.4
idna==2.8
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
requests==2.21.0
urllib3==1.24.2
waitress==1.1.0
whitenoise==3.3.1
